+++ tests/integ/datasource/test_databricks.py
@@ -2,18 +2,19 @@
 # Copyright (c) 2012-2025 Snowflake Computing Inc. All rights reserved.
 #
 
-import datetime
-from decimal import Decimal
+import json
 
 import pytest
 
 from snowflake.snowpark import Row
+from snowflake.snowpark._internal.data_source import DataSourcePartitioner
+from snowflake.snowpark._internal.data_source.drivers.databricks_driver import (
+    DatabricksDriver,
+)
+from snowflake.snowpark._internal.data_source.utils import DBMS_TYPE
 from snowflake.snowpark._internal.utils import (
     random_name_for_temp_object,
     TempObjectType,
-)
-from snowflake.snowpark._internal.data_source.drivers.databricks_driver import (
-    DatabricksDriver,
 )
 from snowflake.snowpark.exceptions import SnowparkDataframeReaderException
 from snowflake.snowpark.types import (
@@ -21,18 +22,17 @@ from snowflake.snowpark.types import (
     StructField,
     LongType,
     StringType,
-    BooleanType,
     BinaryType,
-    DateType,
-    TimestampType,
-    DecimalType,
-    DoubleType,
-    TimestampTimeZone,
     VariantType,
     MapType,
 )
-
 from tests.parameters import DATABRICKS_CONNECTION_PARAMETERS
+from tests.resources.test_data_source_dir.test_databricks_data import (
+    EXPECTED_TEST_DATA,
+    EXPECTED_TYPE,
+    TEST_TABLE_NAME,
+    DATABRICKS_TEST_EXTERNAL_ACCESS_INTEGRATION,
+)
 from tests.utils import IS_IN_STORED_PROC
 
 DEPENDENCIES_PACKAGE_UNAVAILABLE = True
@@ -50,144 +50,6 @@ pytestmark = [
 ]
 
 
-TEST_TABLE_NAME = "ALL_TYPE_TABLE"
-EXPECTED_TEST_DATA = [
-    (
-        -34,
-        25393,
-        35234,
-        5644171805,
-        18.264881134033203,
-        9187.446999674603,
-        Decimal("269.89"),
-        "str_8541",
-        True,
-        bytearray(b"\xad\xa9\xdd\xa2"),
-        datetime.date(2025, 6, 8),
-        datetime.datetime(2025, 4, 16, 17, 39, 39, 565000),
-        datetime.datetime(2025, 4, 16, 17, 49, 8, 565000),
-        "[\n  82,\n  40\n]",
-        '{\n  "key1": 71,\n  "key2": 81\n}',
-        '{\n  "field1": "f_25",\n  "field2": 25\n}',
-        "3-10",
-        "18 14:29:08.000000000",
-    ),
-    (
-        -113,
-        -14623,
-        74665,
-        7633120703,
-        41.216453552246094,
-        1063.0827475381134,
-        Decimal("9960.28"),
-        "str_7962",
-        True,
-        bytearray(b"\xad\xa9\xdd\xa2"),
-        datetime.date(2025, 4, 27),
-        datetime.datetime(2025, 4, 16, 17, 42, 48, 565000),
-        datetime.datetime(2025, 4, 16, 17, 51, 18, 565000),
-        "[\n  6,\n  87\n]",
-        '{\n  "key1": 83,\n  "key2": 12\n}',
-        '{\n  "field1": "f_77",\n  "field2": 13\n}',
-        "1-8",
-        "0 23:14:09.000000000",
-    ),
-    (
-        -96,
-        -431,
-        78281,
-        2300077013,
-        84.57820892333984,
-        8415.70918243513,
-        Decimal("1669.93"),
-        "str_8208",
-        False,
-        bytearray(b"\xad\xa9\xdd\xa2"),
-        datetime.date(2025, 7, 14),
-        datetime.datetime(2025, 4, 16, 17, 40, 11, 565000),
-        datetime.datetime(2025, 4, 16, 17, 42, 39, 565000),
-        "[\n  0,\n  89\n]",
-        '{\n  "key1": 97,\n  "key2": 33\n}',
-        '{\n  "field1": "f_84",\n  "field2": 1\n}',
-        "1-10",
-        "2 11:12:05.000000000",
-    ),
-    (
-        114,
-        11139,
-        75014,
-        1135763646,
-        14.668656349182129,
-        1378.8325065107654,
-        Decimal("7411.91"),
-        "str_9765",
-        False,
-        bytearray(b"\xad\xa9\xdd\xa2"),
-        datetime.date(2025, 6, 29),
-        datetime.datetime(2025, 4, 16, 17, 48, 27, 565000),
-        datetime.datetime(2025, 4, 16, 17, 50, 8, 565000),
-        "[\n  92,\n  27\n]",
-        '{\n  "key1": 52,\n  "key2": 65\n}',
-        '{\n  "field1": "f_85",\n  "field2": 50\n}',
-        "7-4",
-        "22 04:52:41.000000000",
-    ),
-    (
-        -31,
-        -15555,
-        64403,
-        668558045,
-        80.87367248535156,
-        1413.5031507161045,
-        Decimal("9620.13"),
-        "str_4635",
-        False,
-        bytearray(b"\xad\xa9\xdd\xa2"),
-        datetime.date(2025, 7, 2),
-        datetime.datetime(2025, 4, 16, 17, 36, 22, 565000),
-        datetime.datetime(2025, 4, 16, 17, 47, 9, 565000),
-        "[\n  81,\n  65\n]",
-        '{\n  "key1": 67,\n  "key2": 88\n}',
-        '{\n  "field1": "f_98",\n  "field2": 69\n}',
-        "0-7",
-        "19 06:25:08.000000000",
-    ),
-]
-EXPECTED_TYPE = StructType(
-    [
-        StructField("COL_BYTE", LongType(), nullable=True),
-        StructField("COL_SHORT", LongType(), nullable=True),
-        StructField("COL_INT", LongType(), nullable=True),
-        StructField("COL_LONG", LongType(), nullable=True),
-        StructField("COL_FLOAT", DoubleType(), nullable=True),
-        StructField("COL_DOUBLE", DoubleType(), nullable=True),
-        StructField("COL_DECIMAL", DecimalType(10, 2), nullable=True),
-        StructField("COL_STRING", StringType(), nullable=True),
-        StructField("COL_BOOLEAN", BooleanType(), nullable=True),
-        StructField("COL_BINARY", BinaryType(), nullable=True),
-        StructField("COL_DATE", DateType(), nullable=True),
-        StructField(
-            "COL_TIMESTAMP",
-            TimestampType(timezone=TimestampTimeZone.NTZ),
-            nullable=True,
-        ),
-        StructField(
-            "COL_TIMESTAMP_NTZ",
-            TimestampType(timezone=TimestampTimeZone.NTZ),
-            nullable=True,
-        ),
-        StructField("COL_ARRAY", VariantType(), nullable=True),
-        StructField("COL_MAP", VariantType(), nullable=True),
-        StructField("COL_STRUCT", VariantType(), nullable=True),
-        StructField("COL_INTERVAL_YEAR_MONTH", StringType(), nullable=True),
-        StructField("COL_INTERVAL_DAY_TIME", StringType(), nullable=True),
-    ]
-)
-DATABRICKS_TEST_EXTERNAL_ACCESS_INTEGRATION = (
-    "snowpark_dbapi_databricks_test_integration"
-)
-
-
 def create_databricks_connection():
     import databricks.sql
 
@@ -196,8 +58,8 @@ def create_databricks_connection():
 
 @pytest.mark.parametrize(
     "input_type, input_value",
-    # [("table", TEST_TABLE_NAME), ("query", f"(SELECT * FROM {TEST_TABLE_NAME})")],
-    [("table", TEST_TABLE_NAME)],
+    [("table", TEST_TABLE_NAME), ("query", f"(SELECT * FROM {TEST_TABLE_NAME})")],
+    # [("table", TEST_TABLE_NAME)],
 )
 def test_basic_databricks(session, input_type, input_value):
     input_dict = {
@@ -205,7 +67,8 @@ def test_basic_databricks(session, input_type, input_value):
     }
     df = session.read.dbapi(create_databricks_connection, **input_dict)
     ret = df.collect()
-    assert ret == EXPECTED_TEST_DATA and df.schema == EXPECTED_TYPE
+    assert ret == EXPECTED_TEST_DATA
+    assert df.schema == EXPECTED_TYPE
 
     table_name = random_name_for_temp_object(TempObjectType.TABLE)
     df.write.save_as_table(table_name, mode="overwrite", table_type="temp")
@@ -241,7 +104,7 @@ def test_unit_data_source_data_to_pandas_df():
     ]
     df = DatabricksDriver.data_source_data_to_pandas_df(data, schema)
     assert df.to_dict(orient="records") == [
-        {"COL1": 1, "COL2": '{"key1": "value1", "key2": "value2"}'}
+        {"COL1": 1, "COL2": [("key1", "value1"), ("key2", "value2")]}
     ]
 
 
@@ -283,3 +146,26 @@ def test_udtf_ingestion_databricks(session, caplog):
         "TEMPORARY  FUNCTION  data_source_udtf_" "" in caplog.text
         and "table(data_source_udtf" in caplog.text
     )
+
+
+def test_unit_udtf_ingestion(session):
+    dbx_driver = DatabricksDriver(create_databricks_connection, DBMS_TYPE.DATABRICKS_DB)
+    udtf_ingestion_class = dbx_driver.udtf_class_builder()
+    udtf_ingestion_instance = udtf_ingestion_class()
+
+    dsp = DataSourcePartitioner(create_databricks_connection, "ALL_TYPE_TABLE")
+    yield_data = udtf_ingestion_instance.process(dsp.partitions[0])
+    for row, expected_row in zip(yield_data, EXPECTED_TEST_DATA):
+        for index, (field, value) in enumerate(zip(EXPECTED_TYPE.fields, row)):
+            if isinstance(field.datatype, VariantType):
+                # Convert ArrayType, MapType, and StructType to JSON
+                if "map" in field.name.lower():
+                    assert json.loads(value) == json.loads(expected_row[index])
+                else:
+                    assert value == json.loads(expected_row[index])
+            elif isinstance(field.datatype, BinaryType):
+                # Convert BinaryType to hex string
+                assert bytearray(bytes.fromhex(value)) == expected_row[index]
+            else:
+                # Keep other types as is
+                assert value == expected_row[index]